Add support to update reserved list in Cloud SQL
