+++ core/src/main/java/google/registry/tools/UpdateReservedListCommand.java
@@ -15,14 +15,17 @@
 package google.registry.tools;
 
 import static com.google.common.base.Preconditions.checkArgument;
+import static google.registry.model.transaction.TransactionManagerFactory.jpaTm;
 import static google.registry.util.ListNamingUtils.convertFilePathToName;
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 import com.beust.jcommander.Parameters;
 import com.google.common.base.Strings;
 import google.registry.model.registry.label.ReservedList;
+import google.registry.model.registry.label.ReservedListDao;
 import google.registry.util.SystemClock;
 import java.nio.file.Files;
+import java.util.List;
 import java.util.Optional;
 
 /** Command to safely update {@link ReservedList} on Datastore. */
@@ -32,18 +35,38 @@ final class UpdateReservedListCommand extends CreateOrUpdateReservedListCommand 
   @Override
   protected void init() throws Exception {
     name = Strings.isNullOrEmpty(name) ? convertFilePathToName(input) : name;
+    // TODO(shicong): Read existing entry from Cloud SQL
     Optional<ReservedList> existing = ReservedList.get(name);
     checkArgument(
         existing.isPresent(), "Could not update reserved list %s because it doesn't exist.", name);
+    boolean shouldPublish =
+        this.shouldPublish == null ? existing.get().getShouldPublish() : this.shouldPublish;
+    List<String> allLines = Files.readAllLines(input, UTF_8);
     ReservedList.Builder updated =
         existing
             .get()
             .asBuilder()
-            .setReservedListMapFromLines(Files.readAllLines(input, UTF_8))
-            .setLastUpdateTime(new SystemClock().nowUtc());
-    if (shouldPublish != null) {
-      updated.setShouldPublish(shouldPublish);
-    }
+            .setReservedListMapFromLines(allLines)
+            .setLastUpdateTime(new SystemClock().nowUtc())
+            .setShouldPublish(shouldPublish);
     stageEntityChange(existing.get(), updated.build());
+    if (alsoCloudSql) {
+      cloudSqlReservedList =
+          google.registry.schema.tld.ReservedList.create(
+              name, shouldPublish, parseToReservationsByLabels(allLines));
+    }
+  }
+
+  @Override
+  void saveToCloudSql() {
+    jpaTm()
+        .transact(
+            () -> {
+              checkArgument(
+                  ReservedListDao.checkExists(cloudSqlReservedList.getName()),
+                  "A reserved list of this name doesn't exist: %s.",
+                  cloudSqlReservedList.getName());
+              ReservedListDao.save(cloudSqlReservedList);
+            });
   }
 }
