+++ core/src/test/java/google/registry/schema/integration/SqlIntegrationTestSuite.java
@@ -28,7 +28,6 @@ import google.registry.persistence.ZonedDateTimeConverterTest;
 import google.registry.schema.cursor.CursorDaoTest;
 import google.registry.schema.tld.PremiumListDaoTest;
 import google.registry.schema.tmch.ClaimsListDaoTest;
-import google.registry.tools.CreateReservedListCommandTest;
 import google.registry.ui.server.registrar.RegistryLockGetActionTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
@@ -49,7 +48,6 @@ import org.junit.runners.Suite.SuiteClasses;
   BloomFilterConverterTest.class,
   ClaimsListDaoTest.class,
   CreateAutoTimestampConverterTest.class,
-  CreateReservedListCommandTest.class,
   CurrencyUnitConverterTest.class,
   CursorDaoTest.class,
   DateTimeConverterTest.class,
